Don't forbid sending addr-reg option if M=O=0
This may be useful in the future to inform the network of the
device's hostname on networks that don't have DHCPv4 or
DHCPv6 servers.

On such networks, setting M=1 or O=1 might confuse clients
by making them think that there is a DHCPv6 server. Those
clients might, for example, refuse to connect to the network,
or time out, if their DHCPv6 requests do not receive a response.