Force refeshes to occur periodically.
This guarantees that refreshes occur every four hours,
regardless of the address's lifetime.

It's not clear whether this is an improvement, though - the network operator can already force refreshes to occur every 4 hours, by setting the address lifetime to 4 hours.

It might be simpler to instead get rid of the 4-hour minimum
and just add text to point out that the operator can control the
address refresh frequency by setting the RA lifetime. If we do
this, we need to do something about static addresses.