Get rid of the 4-hour minimum.
This didn't really make sense because it did not guarantee that
refreshes were sent every 4 hours. Just remove it, and instead,
note in the text that the network administrator can control the
frequency of refreshes by setting the PIO lifetime appropriately.

Retain the 4-hour value for static addresses, but say that it
SHOULD be configurable.